+++ hashtable/chaining/hashmap_test.go
@@ -1,4 +1,4 @@
-package hashtable
+package chaining
 
 import (
 	"testing"
